Add interactive property and default it to false
The property controls the (inverted) non_interactive option of
Net::SSH. This, in turn, controls whether a password prompt may be
shown by Net::SSH.

In general, I think it is not desirable to have a run get stuck on
a password prompt. Instead, it should fail. In cases where a prompt
is in fact desired, `interactive` can be set to true.